Added ignoreMissing validation option
