+++ riskified-sdk/src/main/java/com/riskified/models/BaseOrder.java
@@ -4,7 +4,10 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
+import com.riskified.validations.FieldBadFormatException;
 import com.riskified.validations.IValidated;
+import com.riskified.validations.Validate;
+import com.riskified.validations.Validation;
 
 public abstract class BaseOrder implements IValidated {
 
@@ -72,6 +75,89 @@ public abstract class BaseOrder implements IValidated {
         noteAttributes = new ArrayList<Attributes>();
         taxLines = new ArrayList<TaxLines>();
     }
+    
+
+	public void validate(Validation validationType)
+			throws FieldBadFormatException {
+		
+		Validate.stringNotNullOrEmpty(this.id, "Id");
+		
+		if(validationType == Validation.all) { // Validated required fields
+			
+			Validate.stringNotNullOrEmpty(this.name, "Name");
+			Validate.stringNotNullOrEmpty(this.email, "Email");
+			Validate.notNull(this.createdAt, "Created At");
+			Validate.notNull(this.closedAt, "Closed At");
+			Validate.notNull(this.updatedAt, "Updated At");
+			Validate.stringNotNullOrEmpty(this.gateway, "Gateway");
+			Validate.stringNotNullOrEmpty(this.browserIp, "Browser IP");
+			Validate.notNull(this.totalPrice, "Total Price");
+			Validate.notNull(this.totalDiscounts, "Total Discounts");
+			Validate.notNull(this.lineItems, "Line Items");
+			Validate.notNull(this.discountCodes, "Discount Codes");
+			Validate.notNull(this.shippingLines, "Shipping Lines");
+			Validate.notNull(this.paymentDetails, "Payment Details");
+			Validate.notNull(this.customer, "Customer");
+			Validate.notNull(this.billingAddress, "Billing Address");
+			Validate.notNull(this.shippingAddress, "Shipping Address");
+		}
+		
+		if(this.totalPrice != null) {
+			Validate.mustBePositive(this.totalPrice, "Total Price");
+		}
+		
+		if(this.browserIp != null) {
+			Validate.ipAddressWellFormed(this.browserIp, "Browser IP");
+		}
+		
+		if(this.currency != null) {
+			Validate.currencyCodeWellFormed(currency, "Currency");
+		}
+		
+		if(this.email != null) {
+			Validate.emailAddressWellFormed(this.email, "Email");
+		}
+		
+		if(this.lineItems != null) {
+			for(LineItem lineItem : this.lineItems) {
+				lineItem.validate(validationType);
+			}
+		}
+		
+		if(this.discountCodes != null) {
+			for(DiscountCode discountCode : this.discountCodes) {
+				discountCode.validate(validationType);
+			}
+		}
+		
+		
+		if(this.shippingLines != null) {
+			for(ShippingLine shippingLine : this.shippingLines) {
+				shippingLine.validate(validationType);
+			}
+		}
+		
+		
+		if(this.paymentDetails != null) {
+			this.paymentDetails.validate(validationType);
+		}
+		
+		
+		if(this.customer != null) {
+			this.customer.validate(validationType);
+		}
+		
+		
+		if(this.billingAddress != null) {
+			this.billingAddress.validate(validationType);
+		}
+		
+		
+		if(this.shippingAddress != null) {
+			this.shippingAddress.validate(validationType);
+		}
+	}
+
