+++ riskified-sdk/src/main/java/com/riskified/validations/Validate.java
@@ -1,6 +1,7 @@
 package com.riskified.validations;
 
 import org.apache.commons.validator.routines.EmailValidator;
+import org.apache.commons.validator.routines.InetAddressValidator;
 
 public class Validate {
 
@@ -18,11 +19,23 @@ public class Validate {
 		}
 	}
 	static public void emailAddressWellFormed(String email, String fieldName) throws FieldBadFormatException {
-		Validate.notNull(email, fieldName);
+		
 		if(!EmailValidator.getInstance().isValid(email)) {
 			throw new FieldBadFormatException("in " + fieldName + " field, value of " + email + " is not a valid email.");
 		}
 		
+	}
+	
+	static public void ipAddressWellFormed(String ip, String fieldName) throws FieldBadFormatException {
+		if(!InetAddressValidator.getInstance().isValidInet4Address(ip)) {
+			throw new FieldBadFormatException("in " + fieldName + " field, value of " + ip + " is not a valid IP address.");
+		}
+	}
+	
+	static public void currencyCodeWellFormed(String currency, String fieldName) throws FieldBadFormatException { 
+		if(currency.length() != 3 || !currency.matches("[A-Z]+")) {
+			throw new FieldBadFormatException("in " + fieldName + " field, value of " + currency + " is not a valid currency code (should be 3 capital letters).");
+		}
 	}
 	
 	static public void mustBePositive(double number, String fieldName) throws FieldBadFormatException {
