+++ riskified-sdk/src/main/java/com/riskified/validations/Validation.java
@@ -8,6 +8,6 @@ package com.riskified.validations;
  */
 public enum Validation {
 	none,
-	weak,
+	ignoreMissing,
 	all
 }
