+++ riskified-sdk/src/main/resources/riskified_sdk.properties
@@ -7,7 +7,7 @@ authKey=ad6b6e6376fb1e3521e44ca28451d58b9605d932
 #debug / sandbox / production
 environment=debug
 
-#none / weak / all
+#none / ignoreMissing / all
 validation=all
 
 #debugRiskifiedHostUrl
